moved safe copy out to it’s own file
