+++ CanvasPlusPlayground/Features/Assignments/UploadSubmissionManager.swift
@@ -206,20 +206,3 @@ public class UploadSubmissionManager {
 enum MimeType: String {
     case txt = "text/plain", other = "application/octet-stream"
 }
-
-extension URL {
-    func safeCopyOut() throws -> URL {
-        let coordinator = NSFileCoordinator()
-        var coordError: NSError?
-        var destURL: URL?
-        let tmp = FileManager.default.temporaryDirectory.appendingPathComponent(self.lastPathComponent)
-
-        coordinator.coordinate(readingItemAt: self, options: [.withoutChanges], error: &coordError) { readURL in
-            try? FileManager.default.removeItem(at: tmp)
-            try! FileManager.default.copyItem(at: readURL, to: tmp)
-            destURL = tmp
-        }
-        if let e = coordError { throw e }
-        return destURL!
-    }
-}
